clean responsive theme

+++ themes/Smoke-N-Wings_Responsive/inc/default.php
@@ -1,10 +1,14 @@
 <?php 
-
 /**
- * Summary of smokeWings_setup_theme default
+ * Summary of smokewings_setup_theme default
+ * 
  * @return void
+ * 
+ * @package smokewings
  */
-function smokeWings_setup_theme()
+
+if ( ! function_exists( "smokewings_setup_theme" ) ) {
+function smokewings_setup_theme()
 {
 
     // Enable post thumbnails for posts and pages
@@ -16,35 +20,40 @@ function smokeWings_setup_theme()
 
 
     register_nav_menus(array(
-        'smokeWings_header_menu' => __('Smoke Wings Header Menu', 'smokeWings'),
-        'smokeWings_mobile_menu' => __('Smoke Wings Mobile Menu', 'smokeWings'),
+        'smokewings_header_menu' => __('Smoke Wings Header Menu', 'smokewings'),
+        'smokewings_mobile_menu' => __('Smoke Wings Mobile Menu', 'smokewings'),
         
         // Footer Menu
-        'smokeWings_footer_menu1' => __('Smoke-N-Wings Footer Menu1', 'smokeWings'),
-        'smokeWings_footer_menu2' => __('Smoke-N-Wings Footer Menu2', 'smokeWings'),
+        'smokewings_footer_menu1' => __('Smoke-N-Wings Footer Menu1', 'smokewings'),
+        'smokewings_footer_menu2' => __('Smoke-N-Wings Footer Menu2', 'smokewings'),
     ));
 }
-add_action('after_setup_theme', 'smokeWings_setup_theme');
+add_action('after_setup_theme', 'smokewings_setup_theme');
+}
 
-function smokeWings_add_time_meta_box() {
+if ( ! function_exists( "smokewings_add_time_meta_box" ) ) {
+    function smokewings_add_time_meta_box() {
     add_meta_box(
         'time_meta_box',                       
         'Set Times',                          
-        'smokeWings_time_meta_box_callback',      
+        'smokewings_time_meta_box_callback',      
         'post',                            
         'side'                             
     );
 }
-add_action('add_meta_boxes', 'smokeWings_add_time_meta_box');
 
-function smokeWings_time_meta_box_callback($post) {
+add_action('add_meta_boxes', 'smokewings_add_time_meta_box');
+}
+if ( ! function_exists( "smokewings_time_meta_box_callback" ) ) {
+    function smokewings_time_meta_box_callback($post) {
 
     $set_time = get_post_meta($post->ID, '_set_time', true);
 
     echo '<input type="text" name="set_time_field" value="' . esc_attr($set_time) . '" style="width:100%;" />';
 }
-
-function smokeWings_save_time_meta_box($post_id) {
+}
+if ( ! function_exists( "smokewings_save_time_meta_box" ) ) {
+    function smokewings_save_time_meta_box($post_id) {
     if ( array_key_exists('set_time_field', $_POST) ) {
         update_post_meta(
             $post_id,
@@ -53,4 +62,5 @@ function smokeWings_save_time_meta_box($post_id) {
         );
     }
 }
-add_action('save_post', 'smokeWings_save_time_meta_box');
+add_action('save_post', 'smokewings_save_time_meta_box');
+}
